requirements.txt
httpx
questionary
tomli
python-dotenv
rich
mmh3
pyyaml